[Chore] UserManager에서 객체 파일 분리
