+++ RunPepperRun/Model/FirestoreCollection.swift
@@ -1,0 +1,13 @@
+//
+//  FirestoreCollection.swift
+//  RunPepperRun
+//
+//  Created by peppermint100 on 1/21/24.
+//
+
+import Foundation
+
+enum FirestoreCollection: String {
+    case user
+    case history
+}
