+++ RunPepperRun/Model/User.swift
@@ -1,0 +1,14 @@
+//
+//  User.swift
+//  RunPepperRun
+//
+//  Created by peppermint100 on 1/21/24.
+//
+
+import Foundation
+
+struct User: Codable {
+    let nickname: String
+    let weight: Double
+}
+
